Always respect disabling of Query/Mutations fields
Strapi parses all plugins for a config/schema.graphql files and merges
those together to build the schema. Where the schema is generated
dynamically (as is the case with our own custom types) then electing
to disable a field as [described in documentation)(https://strapi.io/documentation/3.0.0-beta.x/plugins/graphql.html#disable-a-query-or-a-type)
works as expected; however, where they are described in-situ as is the
case with Users/Roles (from users-permissions plugin) and UploadFiles
(from upload plugin), those configurations don't work as expected.

Removal of field definitions as early as possible in the
chain ensures that all bases are covered. For example, the schema is
used separately to export local cache and to build our federated schema
definition, so it's important the definitions are correct and the same.